+++ scripts/formapplications/item-pile-config.js
@@ -1,8 +1,8 @@
 import CONSTANTS from "../constants.js";
 import API from "../api.js";
-import { ItemPileCurrenciesEditor } from "./itemPileCurrenciesEditor.js";
+import { ItemPileCurrenciesEditor } from "./item-pile-currencies-editor.js";
 import * as lib from "../lib/lib.js";
-import { ItemPileFiltersEditor } from "./itemPileFiltersEditor.js";
+import { ItemPileFiltersEditor } from "./item-pile-filters-editor.js";
 
 export class ItemPileConfig extends FormApplication {
 
@@ -72,7 +72,7 @@ export class ItemPileConfig extends FormApplication {
                 if(isLinked){
                     const doContinue = await Dialog.confirm({
                         title: game.i18n.localize("ITEM-PILES.Dialogs.LinkedActorWarning.Title"),
-                        content: lib.dialogWarning(game.i18n.localize("ITEM-PILES.Dialogs.LinkedActorWarning.Content")),
+                        content: lib.dialogLayout({ message: game.i18n.localize("ITEM-PILES.Dialogs.LinkedActorWarning.Content") }),
                         defaultYes: false
                     });
                     if (!doContinue) {
@@ -199,7 +199,7 @@ export class ItemPileConfig extends FormApplication {
     async resetSharingData(){
         return new Dialog({
             title: game.i18n.localize("ITEM-PILES.Dialogs.ResetSharingData.Title"),
-            content: lib.dialogWarning(game.i18n.localize("ITEM-PILES.Dialogs.ResetSharingData.Content")),
+            content: lib.dialogLayout({ message: game.i18n.localize("ITEM-PILES.Dialogs.ResetSharingData.Content") }),
             buttons: {
                 confirm: {
                     icon: '<i class="fas fa-check"></i>',